requirements.txt
matplotlib>=3.3.4
numpy>=1.20.1
tqdm>=4.59.0
scikit-image>=0.18.1
pillow>=8.1.2
